Minor fixes.

+++ Source/rendering/graphic.cpp
@@ -82,7 +82,7 @@ void VertAttribObject::textureCoordAttrib() {
 
 /*----------------------------------------------------------------------------------*/
 
-void screenColour(float r, float g, float b, float a) {
+void screenColour(double r, double g, double b, double a) {
 	glClearColor(r, g, b, a);
 	glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT);
 }
