+++ Source/rendering/graphic.h
@@ -36,6 +36,6 @@ public:
 	void textureCoordAttrib();
 };
 
-void screenColour(float r, float g, float b, float a);
+void screenColour(double r, double g, double b, double a);
 
 glm::mat4 getMat4Model(unsigned int i, glm::vec3 cube_positions[]);
